microservices_borrowing/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework>=3.12
requests>=2.26